+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/provider/foundationdb/synchronizedsession/SynchronizedSessionRunner.java
@@ -290,6 +290,16 @@ public class SynchronizedSessionRunner implements FDBDatabaseRunner {
     }
 
     @Override
+    public void setTransactionTimeoutMillis(long transactionTimeoutMillis) {
+        underlying.setTransactionTimeoutMillis(transactionTimeoutMillis);
+    }
+
+    @Override
+    public long getTransactionTimeoutMillis() {
+        return underlying.getTransactionTimeoutMillis();
+    }
+
+    @Override
     @Nonnull
     public FDBRecordContext openContext() {
         return underlying.openContext();